extract filters to separate module
